+++ src/FluentDbTools/Tests/Test.FluentDbTools.DbProvider/DbProviderTests.cs
@@ -1,6 +1,13 @@
 using System.Threading.Tasks;
 using FluentDbTools.Common.Abstractions;
 using Example.FluentDbTools.Database;
+using Example.FluentDbTools.Migration;
+using FluentDbTools.Migration;
+using FluentDbTools.Migration.Abstractions;
+using FluentMigrator.Runner;
+using FluentMigrator.Runner.Processors.Oracle;
+using FluentMigrator.Runner.VersionTableInfo;
+using Microsoft.Extensions.DependencyInjection;
 using TestUtilities.FluentDbTools;
 using Xunit;
 
@@ -13,10 +20,20 @@ namespace Test.FluentDbTools.DbProvider
         [MemberData(nameof(TestParameters.DbParameters), MemberType = typeof(TestParameters))]
         public async Task DbProvider_ExampleRepository_Success(SupportedDatabaseTypes databaseType)
         {
+
             var overrideConfig = OverrideConfig.GetInMemoryOverrideConfig(databaseType);
             overrideConfig.Add("database:schemaprefix:id", "EX");
             overrideConfig.Add("database:migration:schemaprefix:id", "EX");
-            await DbExampleExecutor.ExecuteDbExample(databaseType, overrideConfig);
+
+            var provider = MigrationBuilder.BuildMigration(databaseType, overrideConfig);
+            using (provider as ServiceProvider)
+            using (var scope = provider.CreateScope())
+            {
+                var migrationRunner = scope.ServiceProvider.GetService<IMigrationRunner>();
+                migrationRunner.MigrateUp();
+                await DbExampleExecutor.ExecuteDbExample(databaseType, overrideConfig);
+                migrationRunner.DropSchema(scope.ServiceProvider.GetRequiredService<IVersionTableMetaData>());
+            }
         }
 
         [Theory]